fixed type mismatch, and cleanup

+++ ear/src/main/java/cz/cvut/fel/ear/dao/ReviewRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    List<Review> findAllByBoardGame_IdIn(List<Integer> integers);
+    List<Review> findAllByBoardGame_IdIn(List<Long> integers);
 }
